# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MemorialSpace API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /maintenance-schedules:
    get:
      operationId: listMaintenanceSchedules
      tags: [maintenance]
      summary: List maintenance schedules
      parameters:
        - { name: organizationId, in: query, schema: { type: integer } }
        - { name: assetId, in: query, schema: { type: integer } }
        - { name: isActive, in: query, schema: { type: boolean } }
      responses:
        "200":
          description: Schedules
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MaintenanceSchedule" }
    post:
      operationId: createMaintenanceSchedule
      tags: [maintenance]
      summary: Create a maintenance schedule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMaintenanceScheduleBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaintenanceSchedule" }
  /maintenance-schedules/{id}:
    get:
      operationId: getMaintenanceSchedule
      tags: [maintenance]
      summary: Get a maintenance schedule
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: organizationId, in: query, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Schedule
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaintenanceSchedule" }
    put:
      operationId: updateMaintenanceSchedule
      tags: [maintenance]
      summary: Update a maintenance schedule
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMaintenanceScheduleBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaintenanceSchedule" }
    delete:
      operationId: deleteMaintenanceSchedule
      tags: [maintenance]
      summary: Delete a maintenance schedule
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: organizationId, in: query, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
  /maintenance-schedules/{id}/generate:
    post:
      operationId: generateMaintenanceWorkOrder
      tags: [maintenance]
      summary: Generate a work order from a schedule and advance its next-due date
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                organizationId: { type: integer }
              required: [organizationId]
      responses:
        "201":
          description: Generated work order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkOrder" }

  # ─── WORK ORDER COMMENTS ─────────────────────────────────────────────────────
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - { name: organizationId, in: query, schema: { type: integer } }
        - { name: status, in: query, schema: { type: string } }
        - { name: categoryId, in: query, schema: { type: integer } }
      responses:
        "200":
          description: Expenses
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Expense" }
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateExpenseBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Expense" }
  /expenses/{id}:
    get:
      operationId: getExpense
      tags: [expenses]
      summary: Get an expense
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: organizationId, in: query, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Expense
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Expense" }
    put:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateExpenseBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Expense" }
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: organizationId, in: query, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
  /expenses/{id}/approve:
    post:
      operationId: approveExpense
      tags: [expenses]
      summary: Approve an expense
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                organizationId: { type: integer }
                approvedBy: { type: ["integer", "null"] }
              required: [organizationId]
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Expense" }
  /expenses/{id}/reject:
    post:
      operationId: rejectExpense
      tags: [expenses]
      summary: Reject an expense
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                organizationId: { type: integer }
                approvedBy: { type: ["integer", "null"] }
                notes: { type: ["string", "null"] }
              required: [organizationId]
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Expense" }
  /expenses/{id}/mark-paid:
    post:
      operationId: markExpensePaid
      tags: [expenses]
      summary: Mark an expense as paid
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                organizationId: { type: integer }
              required: [organizationId]
      responses:
        "200":
          description: Paid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Expense" }
components:
  schemas:
...
    WorkOrder:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
        plotId:
          type: ["integer", "null"]
        assetId:
          type: ["integer", "null"]
        assignedTo:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [maintenance, burial, cleaning, inspection, other]
        status:
          type: string
          enum: [open, in_progress, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        dueDate:
          type: ["string", "null"]
          format: date
        completedAt:
          type: ["string", "null"]
          format: date-time
        laborHours:
          type: ["string", "null"]
        laborCost:
          type: ["string", "null"]
        materialsCost:
          type: ["string", "null"]
        completionNotes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - organizationId
        - title
        - type
        - status
        - priority
        - createdAt
        - updatedAt
...
    MaintenanceSchedule:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
        assetId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        frequency:
          type: string
          enum: [daily, weekly, monthly, quarterly, yearly, custom]
        intervalDays:
          type: integer
        lastPerformedAt:
          type: ["string", "null"]
          format: date
        nextDueAt:
          type: ["string", "null"]
          format: date
        assignedTo:
          type: ["integer", "null"]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, organizationId, title, frequency, intervalDays, priority, isActive, createdAt, updatedAt]
    CreateMaintenanceScheduleBody:
      type: object
      properties:
        organizationId:
          type: integer
        assetId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        frequency:
          type: string
          enum: [daily, weekly, monthly, quarterly, yearly, custom]
        intervalDays:
          type: integer
        lastPerformedAt:
          type: ["string", "null"]
          format: date
        nextDueAt:
          type: ["string", "null"]
          format: date
        assignedTo:
          type: ["integer", "null"]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        isActive:
          type: boolean
      required: [organizationId, title, frequency, intervalDays]
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
        categoryId:
          type: ["integer", "null"]
        workOrderId:
          type: ["integer", "null"]
        vendorName:
          type: ["string", "null"]
        description:
          type: string
        amount:
          type: string
        expenseDate:
          type: string
          format: date
        paymentMethod:
          type: string
          enum: [cash, card, check, transfer, other]
        receiptUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected, paid]
        submittedBy:
          type: ["integer", "null"]
        approvedBy:
          type: ["integer", "null"]
        approvedAt:
          type: ["string", "null"]
          format: date-time
        paidAt:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, organizationId, description, amount, expenseDate, paymentMethod, status, createdAt, updatedAt]
    CreateExpenseBody:
      type: object
      properties:
        organizationId:
          type: integer
        categoryId:
          type: ["integer", "null"]
        workOrderId:
          type: ["integer", "null"]
        vendorName:
          type: ["string", "null"]
        description:
          type: string
        amount:
          type: string
        expenseDate:
          type: string
          format: date
        paymentMethod:
          type: string
          enum: [cash, card, check, transfer, other]
        receiptUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected, paid]
        submittedBy:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
      required: [organizationId, description, amount, expenseDate, paymentMethod]
...